requirements.txt
Django==3.2.16
djangorestframework==3.14.0
psycopg2==2.9.7
celery[redis]==5.3.6
celery_singleton==0.3.1
flower
django-redis==5.2.0
django-debug-toolbar==3.0.0
djoser==2.1.0
coverage==7.6.10
